+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.1.1 (TBD)
 
+* Feature Overlay Query message builder, check if features are indexed first
 * TBD
 
 ## [1.1.0](https://github.com/ngageoint/geopackage-ios/releases/tag/1.1.0) (11-04-2015)